fix: change on BIGINT delta

+++ internal/server/db/migrations/000001_create_metrics_table.up.sql
@@ -3,7 +3,7 @@ BEGIN TRANSACTION;
 CREATE TABLE IF NOT EXISTS metrics(
    id varchar PRIMARY KEY,
    type VARCHAR (50) NOT NULL,
-   delta INTEGER NULL,
+   delta BIGINT NULL,
    value DOUBLE PRECISION NULL
 );
